Consolidate dig command
