--- inodaqv2/command_dig.h
@@ -1,8 +1,0 @@
-#pragma once
-
-#include "Arduino.h"
-
-namespace Command
-{
-void command_dig(const ::String &command);
-}
